Move anomaly detection to a separate module
